requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiounittest==1.3.0
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.7.0
binaryornot==0.4.4
botbuilder-core==4.14.2
botbuilder-dialogs
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
certifi==2022.5.18.1
cffi==1.15.0
chardet
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cookiecutter==2.1.1
cryptography==3.4.8
datedelta==1.4
dateparser
emoji==1.7.0
frozenlist==1.3.0
Flask
googletrans==3.1.0a0
grapheme==0.6.0
gspread
idna==2.10
isodate==0.6.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
jsonpickle==1.4.2
MarkupSafe==2.1.1
msal==1.6.0
msrest==0.6.19
multidict==6.0.2
multipledispatch==0.6.0
numpy==1.22.4
oauthlib==3.2.0
parsedatetime==2.6
Pillow==9.1.1
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
recognizers-text==1.0.2a2
recognizers-text-suite==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
recognizers-text-sequence==1.0.2a2
regex
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
text-unidecode==1.3
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
word2number
yarl==1.7.2